Fix D1 import resume paths, pgloader validation, and JSON errors.
Validate per-table pgloader loads against staged SQLite row counts, normalize input paths for resume, route start/verify client failures through the JSON envelope, and consolidate related tests and pgloader helpers.
