+++ internal/import/d1/convert.go
@@ -6,6 +6,8 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+
+	"github.com/planetscale/cli/internal/postgres"
 )
 
 var (
@@ -101,7 +103,7 @@ func ConvertSchema(inputPath, outputPath string) (int, error) {
 
 func convertTableDDL(table TableSchema, all []TableSchema, ctx *TypeCoercionContext) string {
 	var b strings.Builder
-	fmt.Fprintf(&b, "CREATE TABLE IF NOT EXISTS %s (\n", quoteIdent(table.Name))
+	fmt.Fprintf(&b, "CREATE TABLE IF NOT EXISTS %s (\n", postgres.QuoteIdentifier(table.Name))
 
 	var lines []string
 	for _, col := range table.Columns {
@@ -122,7 +124,7 @@ func convertColumn(col ColumnSchema, table TableSchema, all []TableSchema, ctx *
 	pgType := sqliteTypeToPostgres(col, table, all, ctx)
 
 	var parts []string
-	parts = append(parts, quoteIdent(col.Name), pgType)
+	parts = append(parts, postgres.QuoteIdentifier(col.Name), pgType)
 
 	if col.AutoIncrement {
 		parts = append(parts, "GENERATED BY DEFAULT AS IDENTITY")
@@ -318,11 +320,6 @@ func sqliteFunctionArg(s, fn string) string {
 		}
 	}
 	return ""
-}
-
-func quoteIdent(name string) string {
-	escaped := strings.ReplaceAll(name, `"`, `""`)
-	return `"` + escaped + `"`
 }
 
 // ConvertCreateStatement converts a raw SQLite CREATE TABLE line to Postgres (for tests).
