+++ internal/import/d1/pgloader_test.go
@@ -7,6 +7,26 @@ import (
 	"testing"
 )
 
+const pgloaderOrganizationsOK = `
+2026-06-29T17:07:37.780572-04:00 LOG report summary reset
+             table name     errors       rows      bytes      total time
+-----------------------  ---------  ---------  ---------  --------------
+                  fetch          0          0                     0.000s
+        fetch meta data          0          1                     0.021s
+               Truncate          0          1                     0.053s
+          organizations          0         28     2.7 kB          0.775s
+      Total import time          ✓         28     2.7 kB          1.770s
+`
+
+const pgloaderTeamMembersZero = `
+2026-06-29T17:10:50.659297-04:00 LOG report summary reset
+             table name     errors       rows      bytes      total time
+-----------------------  ---------  ---------  ---------  --------------
+                  fetch          0          0                     0.000s
+        fetch meta data          0          0                     0.012s
+      Total import time          ✓          0                     0.966s
+`
+
 func TestBuildPgloaderScriptDataOnlyPerTable(t *testing.T) {
 	table := TableSchema{
 		Name: "organizations",
@@ -145,6 +165,92 @@ func TestSQLLikeMatch(t *testing.T) {
 	}
 	if sqlLikeMatch("user_data", "users_data") {
 		t.Fatal("expected user_data pattern not to match users_data")
+	}
+}
+
+func TestPgloaderHadErrors(t *testing.T) {
+	tests := []struct {
+		name   string
+		output string
+		want   bool
+	}{
+		{
+			name: "clean summary",
+			output: `
+| errors | rows | bytes | total time
+|      0 |  100 |  1 kB | 1.000 s
+`,
+			want: false,
+		},
+		{
+			name: "summary with errors",
+			output: `
+| errors | rows | bytes | total time
+|      3 |   97 |  1 kB | 1.000 s
+`,
+			want: true,
+		},
+		{
+			name:   "database error",
+			output: "Database error 42501: must be owner of table users",
+			want:   true,
+		},
+		{
+			name:   "insufficient privilege",
+			output: "INSUFFICIENT-PRIVILEGE disable triggers",
+			want:   true,
+		},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			if got := pgloaderHadErrors(tt.output); got != tt.want {
+				t.Fatalf("pgloaderHadErrors() = %v, want %v", got, tt.want)
+			}
+		})
+	}
+}
+
+func TestPgloaderFetchMetaDataTableCount(t *testing.T) {
+	if got := pgloaderFetchMetaDataTableCount(pgloaderOrganizationsOK); got != 1 {
+		t.Fatalf("organizations meta = %d, want 1", got)
+	}
+	if got := pgloaderFetchMetaDataTableCount(pgloaderTeamMembersZero); got != 0 {
+		t.Fatalf("team_members meta = %d, want 0", got)
+	}
+	if got := pgloaderFetchMetaDataTableCount("no summary"); got != -1 {
+		t.Fatalf("missing meta = %d, want -1", got)
+	}
+}
+
+func TestPgloaderRowsCopied(t *testing.T) {
+	rows, ok := pgloaderRowsCopied(pgloaderOrganizationsOK, "organizations")
+	if !ok || rows != 28 {
+		t.Fatalf("organizations rows = (%d, %v), want (28, true)", rows, ok)
+	}
+	rows, ok = pgloaderRowsCopied(pgloaderTeamMembersZero, "team_members")
+	if ok || rows != 0 {
+		t.Fatalf("team_members rows = (%d, %v), want (0, false)", rows, ok)
+	}
+}
+
+func TestValidatePgloaderTableLoad(t *testing.T) {
+	if err := validatePgloaderTableLoad(pgloaderOrganizationsOK, "organizations", 28); err != nil {
+		t.Fatalf("expected ok load: %v", err)
+	}
+	if err := validatePgloaderTableLoad(pgloaderOrganizationsOK, "organizations", 0); err == nil {
+		t.Fatal("expected error when staged SQLite row count does not match pgloader output")
+	}
+	if err := validatePgloaderTableLoad(pgloaderOrganizationsOK, "organizations", 30); err == nil {
+		t.Fatal("expected error for row count mismatch")
+	}
+	if err := validatePgloaderTableLoad(pgloaderTeamMembersZero, "team_members", 700); err == nil {
+		t.Fatal("expected error for 0-row load")
+	} else if me, ok := err.(*MigrationError); !ok || me.Info.Code != ErrCodeImportFailed {
+		t.Fatalf("error = %#v", err)
+	}
+	if err := validatePgloaderTableLoad(pgloaderTeamMembersZero, "team_members", 0); err == nil {
+		t.Fatal("expected error when pgloader matched 0 source tables")
 	}
 }
